refactor(e2e): clean up settings test suite
- Sets "waitFor" timeout globally
- Extracts some functions to helper files
- Generally cleans up the test suite
