+++ test/e2e/helpers/actions.ts
@@ -1,4 +1,12 @@
 import type { Workbench } from 'wdio-vscode-service'
+import { Key } from 'webdriverio'
+
+export async function clearInput(input: WebdriverIO.Element) {
+  // Focus the input, sometimes clicking doesn't work
+  await input.setValue('')
+  await browser.keys([Key.Ctrl, 'a'])
+  await browser.keys(Key.Backspace)
+}
 
 /**
  * Opens the Radicle view container in the sidebar, by clicking the radicle button in the
