+++ test/e2e/helpers/assertions.ts
@@ -1,5 +1,25 @@
 import { browser } from '@wdio/globals'
-import type { Workbench } from 'wdio-vscode-service'
+import type { OutputView, Workbench } from 'wdio-vscode-service'
+
+/**
+ * Asserts the output view contains the expected text.
+ */
+export async function expectOutputToContain(outputView: OutputView, expected: string) {
+  await browser.waitUntil(
+    async () => {
+      /**
+       * The text in the output console is split by newlines, which can be affected by the size
+       * of the window. To avoid this, we join the text into a single string.
+       */
+      const joinedText = (await outputView.getText()).join('')
+
+      return joinedText.includes(expected)
+    },
+    {
+      timeoutMsg: `expected the output text to contain "${expected}"`,
+    },
+  )
+}
 
 /**
  * Asserts that the CLI Commands and Patches sections are visible in the sidebar. This is
@@ -23,11 +43,6 @@ export async function expectStandardSidebarViewsToBeVisible(workbench: Workbench
         return false
       }
     },
-    {
-      timeoutMsg: 'expected the standard sidebar views to be visible',
-      // TODO: zac fine tune these (globally?)
-      timeout: 20000,
-      interval: 500,
-    },
+    { timeoutMsg: 'expected the standard sidebar views to be visible' },
   )
 }
